+++ src/rust/engine/Cargo.toml
@@ -19,15 +19,12 @@ cc = "1.0"
 # [workspace]
 
 [dependencies]
-blake2-rfc = "0.2.17"
+boxfuture = { path = "boxfuture" }
 fnv = "1.0.5"
+fs = { path = "fs" }
 futures = "0.1.16"
 futures-cpupool = "0.1.6"
-glob = "0.2.11"
-ignore = "0.1.8"
 lazy_static = "0.2.2"
 ordermap = "0.2.8"
 petgraph = "0.4.5"
 process_executor = { path = "process_executor" }
-tar = "0.4.13"
-tempdir = "0.3.5"
