+++ src/rust/engine/boxfuture/.gitignore
@@ -1,0 +1,1 @@
+Cargo.lock # See http://doc.crates.io/faq.html#why-do-binaries-have-cargolock-in-version-control-but-not-libraries
