+++ src/rust/engine/boxfuture/src/lib.rs
@@ -2,6 +2,8 @@
 // https://github.com/alexcrichton/futures-rs/issues/228 has background for its removal.
 // This avoids needing to call Box::new() around every future that we produce.
 
+extern crate futures;
+
 use futures::future::Future;
 
 pub type BoxFuture<T, E> = Box<Future<Item = T, Error = E> + Send>;
