+++ src/rust/engine/fs/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+version = "0.0.1"
+name = "fs"
+authors = [ "Pants Build <pantsbuild@gmail.com>" ]
+
+[dependencies]
+boxfuture = { path = "../boxfuture" }
+blake2-rfc = "0.2.17"
+futures = "0.1.16"
+futures-cpupool = "0.1.6"
+glob = "0.2.11"
+ignore = "0.1.8"
+lazy_static = "0.2.2"
+ordermap = "0.2.8"
+tar = "0.4.13"
+tempdir = "0.3.5"
